refactor(stm): move read-only IVC asset layer out of the test tree for benchmark reuse
