+++ mithril-stm/src/circuits/halo2_ivc/mod.rs
@@ -55,6 +55,9 @@ pub(crate) mod certificate_proof;
 pub(crate) mod circuit;
 pub(crate) mod config;
 pub(crate) mod constraint_builder;
+#[cfg(any(test, feature = "benchmark-internals"))]
+#[cfg_attr(not(test), allow(dead_code))]
+pub(crate) mod embedded_assets;
 pub(crate) mod errors;
 pub(crate) mod gadgets;
 pub(crate) mod io;
